Telegram-bot/requirements.txt
pyTelegramBotAPI==3.6.5
requests==2.19.1
python-dotenv==0.9.1
